requirements.txt
beautifulsoup4==4.5.1
cssmin==0.2.0
dj-database-url==0.4.1
Django==1.10.2
django-blog-zinnia==0.18.1
django-contact-form==1.2
django-contrib-comments==1.7.3
django-debug-toolbar==1.6
django-guardian==1.4.6
django-material==0.11.0
django-mptt==0.8.7
django-tagging==0.4.5
django-threadedcomments==1.0.1
django-userena==2.0.1
django-xmlrpc==0.1.7
easy-thumbnails==2.3
gunicorn==19.6.0
html2text==2014.12.29
Jinja2==2.8
MarkupSafe==0.23
mots-vides==2015.5.11
Pillow==3.4.1
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
regex==2016.11.21
render==1.0.0
six==1.10.0
sqlparse==0.2.2
virtualenv==15.0.3
virtualenvwrapper-win==1.2.1
whitenoise==3.2.2
